+++ app/Http/Controllers/AchievementController.php
@@ -1,12 +1,10 @@
 <?php
 
-namespace App\Http\Controllers\Api;
+namespace App\Http\Controllers;
 
-use App\Http\Controllers\Controller;
 use App\Http\Requests\Achievement\StoreRequest;
 use App\Http\Requests\Achievement\UpdateRequest;
 use App\Models\Achievement;
-use Illuminate\Http\Request;
 
 class AchievementController extends Controller
 {
