+++ app/Http/Controllers/UploadFileController.php
@@ -1,8 +1,7 @@
 <?php
 
-namespace App\Http\Controllers\Api;
+namespace App\Http\Controllers;
 
-use App\Http\Controllers\Controller;
 use App\Http\Requests\UploadFileRequest;
 
 class UploadFileController extends Controller
